api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
anthropic==0.26.0
structlog==24.1.0
python-multipart==0.0.9
httpx==0.27.0
python-ulid==2.7.0
psycopg2-binary==2.9.9
